Fixed bug in unescape/repositioning logic
We now explicitly track the last match end in its own variable, since in the general case neither the original nor the extracted offset will necessarily match the correct "char" position in Java terms - an HTML entity advances origOffset in a way that matches Java but extractedOffset falls behind, supplementary characters advance extractedOffset in line with Java but origOffset falls behind.  The previous logic was creating additional spurious repositioning records in the presence of supplementary characters, which in turn caused some of the entities to be off by a few positions.